renames playgroundxyz bid adapter to pxyz
